Update subnet edge vlan tags.
In this implementation, provisioning in a subnet (say I2 DCS net) is contrained
by both VLAN and TDM timeslot constaints at edge. While the TDM constraints are
posed on the intra-domain edge ports/links,  the VLAN constraints have to be posed
on the inter-domain links that represent the intra-domain subnet edge ports.
Both NARB and RCE have to recognize patterns for this kind of provisioning and
update vlan tags and bandwidth on the inter-domain links that correspond to
the provisioned subnet edge ports.

Note that we may also add extra information on the intra-domain edge to represent
VLAN tag utilization, which could eleminate the needs for configuring an inter-domain
topology. However, that addition is considered technically more difficult and
will not be put into implementation at this point.


git-svn-id: file:///var/www/svn/narb-sw/trunk@747 aacf9c1e-ac80-435b-8e83-6affbfa3956d