Fixed a bug:
The output of the 'show topology' command in the NARB's RCE CLI can be misleading.  There are two possible commands: 'show topology intradomain' and 'show topology interdomain' If the user issues the command "show topology intr", the *intra* domain topology is NOT shown, instead the *inter* domain topology is displayed.

Several other commands with similar issues are also fixed.


git-svn-id: file:///var/www/svn/narb-sw/trunk@847 aacf9c1e-ac80-435b-8e83-6affbfa3956d