Patch to eliminate lots of compiler warning messages like this when building under gcc/g++-4.3:
 warning: deprecated conversion from string constant to 'char*'

Added (char*) to all of the quoted literals that the compiler was complaining about.

In addition, #include <stdlib.h> is added to a number of files since gcc/g++-4.3
seems to complain about not being able to find declarations for functions like
atoi(), free(), and many others.  (This did not seem to be a problem in older
build environments...)

-chris tracy (chris@maxgigapop.net)


git-svn-id: file:///var/www/svn/narb-sw/trunk@1012 aacf9c1e-ac80-435b-8e83-6affbfa3956d